+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.33
+
+* changed Colors name from custom*** to hemicycle***
+(i.e. customVoteFor (green) is now hemicycleVoteFor)
+
 ## 0.0.30
 
 * Updating informations for pub.dev scoring...